+++ src/commands/tokens/send_ft/amount_ft.rs
@@ -1,0 +1,207 @@
+use std::str::FromStr;
+
+use inquire::Text;
+use serde_json::json;
+
+#[derive(Debug, Clone, interactive_clap::InteractiveClap)]
+#[interactive_clap(input_context = super::SendFtCommandContext)]
+#[interactive_clap(output_context = AmountFtContext)]
+pub struct AmountFt {
+    /// Enter an amount FT to transfer:
+    amount: u128,
+    #[interactive_clap(named_arg)]
+    /// Enter gas for function call
+    prepaid_gas: PrepaidGas,
+}
+
+#[derive(Debug, Clone)]
+pub struct AmountFtContext {
+    global_context: crate::GlobalContext,
+    signer_account_id: near_primitives::types::AccountId,
+    ft_contract_account_id: near_primitives::types::AccountId,
+    receiver_account_id: near_primitives::types::AccountId,
+    amount: u128,
+}
+
+impl AmountFtContext {
+    pub fn from_previous_context(
+        previous_context: super::SendFtCommandContext,
+        scope: &<AmountFt as interactive_clap::ToInteractiveClapContextScope>::InteractiveClapContextScope,
+    ) -> color_eyre::eyre::Result<Self> {
+        Ok(Self {
+            global_context: previous_context.global_context,
+            signer_account_id: previous_context.signer_account_id,
+            ft_contract_account_id: previous_context.ft_contract_account_id,
+            receiver_account_id: previous_context.receiver_account_id,
+            amount: scope.amount,
+        })
+    }
+}
+
+#[derive(Debug, Clone, interactive_clap::InteractiveClap)]
+#[interactive_clap(input_context = AmountFtContext)]
+#[interactive_clap(output_context = PrepaidGasContext)]
+pub struct PrepaidGas {
+    #[interactive_clap(skip_default_input_arg)]
+    /// Enter gas for function call:
+    gas: crate::common::NearGas,
+    #[interactive_clap(named_arg)]
+    /// Enter deposit for a function call
+    attached_deposit: Deposit,
+}
+
+#[derive(Debug, Clone)]
+pub struct PrepaidGasContext {
+    global_context: crate::GlobalContext,
+    signer_account_id: near_primitives::types::AccountId,
+    ft_contract_account_id: near_primitives::types::AccountId,
+    receiver_account_id: near_primitives::types::AccountId,
+    amount: u128,
+    gas: crate::common::NearGas,
+}
+
+impl PrepaidGasContext {
+    pub fn from_previous_context(
+        previous_context: AmountFtContext,
+        scope: &<PrepaidGas as interactive_clap::ToInteractiveClapContextScope>::InteractiveClapContextScope,
+    ) -> color_eyre::eyre::Result<Self> {
+        Ok(Self {
+            global_context: previous_context.global_context,
+            signer_account_id: previous_context.signer_account_id,
+            ft_contract_account_id: previous_context.ft_contract_account_id,
+            receiver_account_id: previous_context.receiver_account_id,
+            amount: previous_context.amount,
+            gas: scope.gas,
+        })
+    }
+}
+
+impl PrepaidGas {
+    fn input_gas(
+        _context: &AmountFtContext,
+    ) -> color_eyre::eyre::Result<Option<crate::common::NearGas>> {
+        eprintln!();
+        let gas = loop {
+            match crate::common::NearGas::from_str(
+                &Text::new("Enter gas for function call:")
+                    .with_initial_value("100 TeraGas")
+                    .prompt()?,
+            ) {
+                Ok(input_gas) => {
+                    if input_gas <= near_gas::NearGas::from_tgas(300) {
+                        break input_gas;
+                    } else {
+                        eprintln!("You need to enter a value of no more than 300 TERAGAS")
+                    }
+                }
+                Err(err) => return Err(color_eyre::Report::msg(err)),
+            }
+        };
+        Ok(Some(gas))
+    }
+}
+
+#[derive(Debug, Clone, interactive_clap::InteractiveClap)]
+#[interactive_clap(input_context = PrepaidGasContext)]
+#[interactive_clap(output_context = DepositContext)]
+pub struct Deposit {
+    #[interactive_clap(skip_default_input_arg)]
+    /// Enter deposit for a function call:
+    deposit: crate::types::near_token::NearToken,
+    #[interactive_clap(named_arg)]
+    /// Select network
+    network_config: crate::network_for_transaction::NetworkForTransactionArgs,
+}
+
+#[derive(Clone)]
+pub struct DepositContext(crate::commands::ActionContext);
+
+impl DepositContext {
+    pub fn from_previous_context(
+        previous_context: PrepaidGasContext,
+        scope: &<Deposit as interactive_clap::ToInteractiveClapContextScope>::InteractiveClapContextScope,
+    ) -> color_eyre::eyre::Result<Self> {
+        let on_after_getting_network_callback: crate::commands::OnAfterGettingNetworkCallback =
+            std::sync::Arc::new({
+                let signer_account_id = previous_context.signer_account_id.clone();
+                let ft_contract_account_id = previous_context.ft_contract_account_id.clone();
+                let receiver_account_id = previous_context.receiver_account_id.clone();
+                let deposit = scope.deposit;
+
+                move |_network_config| {
+                    Ok(crate::commands::PrepopulatedTransaction {
+                        signer_id: signer_account_id.clone(),
+                        receiver_id: ft_contract_account_id.clone(),
+                        actions: vec![near_primitives::transaction::Action::FunctionCall(
+                            near_primitives::transaction::FunctionCallAction {
+                                method_name: "ft_transfer".to_string(),
+                                args: serde_json::to_vec(&json!({
+                                    "receiver_id": receiver_account_id.to_string(),
+                                    "amount": previous_context.amount.to_string()
+                                }))?,
+                                gas: previous_context.gas.as_gas(),
+                                deposit: deposit.as_yoctonear(),
+                            },
+                        )],
+                    })
+                }
+            });
+
+        let on_after_sending_transaction_callback: crate::transaction_signature_options::OnAfterSendingTransactionCallback = std::sync::Arc::new({
+            let signer_account_id = previous_context.signer_account_id.clone();
+            let amount = previous_context.amount;
+            let ft_contract_account_id = previous_context.ft_contract_account_id.clone();
+            let receiver_account_id = previous_context.receiver_account_id.clone();
+
+            move |outcome_view, _network_config| {
+                if let near_primitives::views::FinalExecutionStatus::SuccessValue(_) = outcome_view.status {
+                    eprintln!(
+                        "<{signer_account_id}> has successfully transferred {amount} FT ({ft_contract_account_id}) to <{receiver_account_id}>.",
+                    );
+                }
+                Ok(())
+            }
+        });
+
+        Ok(Self(crate::commands::ActionContext {
+            global_context: previous_context.global_context,
+            interacting_with_account_ids: vec![
+                previous_context.ft_contract_account_id,
+                previous_context.signer_account_id,
+                previous_context.receiver_account_id,
+            ],
+            on_after_getting_network_callback,
+            on_before_signing_callback: std::sync::Arc::new(
+                |_prepolulated_unsinged_transaction, _network_config| Ok(()),
+            ),
+            on_before_sending_transaction_callback: std::sync::Arc::new(
+                |_signed_transaction, _network_config, _message| Ok(()),
+            ),
+            on_after_sending_transaction_callback,
+        }))
+    }
+}
+
+impl From<DepositContext> for crate::commands::ActionContext {
+    fn from(item: DepositContext) -> Self {
+        item.0
+    }
+}
+
+impl Deposit {
+    fn input_deposit(
+        _context: &PrepaidGasContext,
+    ) -> color_eyre::eyre::Result<Option<crate::types::near_token::NearToken>> {
+        eprintln!();
+        match crate::types::near_token::NearToken::from_str(
+            &Text::new(
+                "Enter deposit for a function call (example: 10NEAR or 0.5near or 10000yoctonear):",
+            )
+            .with_initial_value("1 yoctoNEAR")
+            .prompt()?,
+        ) {
+            Ok(deposit) => Ok(Some(deposit)),
+            Err(err) => Err(color_eyre::Report::msg(err)),
+        }
+    }
+}
