+++ src/commands/tokens/send_ft/mod.rs
@@ -1,7 +1,4 @@
-use std::str::FromStr;
-
-use inquire::Text;
-use serde_json::json;
+mod amount_ft;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
 #[interactive_clap(input_context = super::TokensCommandsContext)]
@@ -13,19 +10,9 @@ pub struct SendFtCommand {
     #[interactive_clap(skip_default_input_arg)]
     /// What is the receiver account ID?
     receiver_account_id: crate::types::account_id::AccountId,
-    /// Enter an amount FT to transfer:
-    amount: u128,
-    #[interactive_clap(long = "prepaid-gas")]
-    #[interactive_clap(skip_default_input_arg)]
-    /// Enter gas for function call:
-    gas: crate::common::NearGas,
-    #[interactive_clap(long = "attached-deposit")]
-    #[interactive_clap(skip_default_input_arg)]
-    /// Enter deposit for a function call:
-    deposit: crate::types::near_token::NearToken,
     #[interactive_clap(named_arg)]
-    /// Select network
-    network_config: crate::network_for_transaction::NetworkForTransactionArgs,
+    /// Specify amount FT
+    amount_ft: self::amount_ft::AmountFt,
 }
 
 #[derive(Debug, Clone)]
@@ -34,9 +21,6 @@ pub struct SendFtCommandContext {
     signer_account_id: near_primitives::types::AccountId,
     ft_contract_account_id: near_primitives::types::AccountId,
     receiver_account_id: near_primitives::types::AccountId,
-    amount: u128,
-    gas: crate::common::NearGas,
-    deposit: crate::types::near_token::NearToken,
 }
 
 impl SendFtCommandContext {
@@ -49,72 +33,7 @@ impl SendFtCommandContext {
             signer_account_id: previous_context.owner_account_id,
             ft_contract_account_id: scope.ft_contract_account_id.clone().into(),
             receiver_account_id: scope.receiver_account_id.clone().into(),
-            amount: scope.amount,
-            gas: scope.gas,
-            deposit: scope.deposit,
         })
-    }
-}
-
-impl From<SendFtCommandContext> for crate::commands::ActionContext {
-    fn from(item: SendFtCommandContext) -> Self {
-        let on_after_getting_network_callback: crate::commands::OnAfterGettingNetworkCallback =
-            std::sync::Arc::new({
-                let signer_account_id = item.signer_account_id.clone();
-                let ft_contract_account_id = item.ft_contract_account_id.clone();
-                let receiver_account_id = item.receiver_account_id.clone();
-
-                move |_network_config| {
-                    Ok(crate::commands::PrepopulatedTransaction {
-                        signer_id: signer_account_id.clone(),
-                        receiver_id: ft_contract_account_id.clone(),
-                        actions: vec![near_primitives::transaction::Action::FunctionCall(
-                            near_primitives::transaction::FunctionCallAction {
-                                method_name: "ft_transfer".to_string(),
-                                args: serde_json::to_vec(&json!({
-                                    "receiver_id": receiver_account_id.to_string(),
-                                    "amount": item.amount.to_string()
-                                }))?,
-                                gas: item.gas.as_gas(),
-                                deposit: item.deposit.as_yoctonear(),
-                            },
-                        )],
-                    })
-                }
-            });
-
-        let on_after_sending_transaction_callback: crate::transaction_signature_options::OnAfterSendingTransactionCallback = std::sync::Arc::new({
-            let signer_account_id = item.signer_account_id.clone();
-            let amount = item.amount;
-            let ft_contract_account_id = item.ft_contract_account_id.clone();
-            let receiver_account_id = item.receiver_account_id.clone();
-
-            move |outcome_view, _network_config| {
-                if let near_primitives::views::FinalExecutionStatus::SuccessValue(_) = outcome_view.status {
-                    eprintln!(
-                        "<{signer_account_id}> has successfully transferred {amount} FT ({ft_contract_account_id}) to <{receiver_account_id}>.",
-                    );
-                }
-                Ok(())
-            }
-        });
-
-        Self {
-            global_context: item.global_context,
-            interacting_with_account_ids: vec![
-                item.ft_contract_account_id,
-                item.signer_account_id,
-                item.receiver_account_id,
-            ],
-            on_after_getting_network_callback,
-            on_before_signing_callback: std::sync::Arc::new(
-                |_prepolulated_unsinged_transaction, _network_config| Ok(()),
-            ),
-            on_before_sending_transaction_callback: std::sync::Arc::new(
-                |_signed_transaction, _network_config, _message| Ok(()),
-            ),
-            on_after_sending_transaction_callback,
-        }
     }
 }
 
@@ -135,44 +54,5 @@ impl SendFtCommand {
             &context.global_context.config.credentials_home_dir,
             "What is the receiver account ID?",
         )
-    }
-
-    fn input_gas(
-        _context: &super::TokensCommandsContext,
-    ) -> color_eyre::eyre::Result<Option<crate::common::NearGas>> {
-        eprintln!();
-        let gas = loop {
-            match crate::common::NearGas::from_str(
-                &Text::new("Enter gas for function call:")
-                    .with_initial_value("100 TeraGas")
-                    .prompt()?,
-            ) {
-                Ok(input_gas) => {
-                    if input_gas <= near_gas::NearGas::from_tgas(300) {
-                        break input_gas;
-                    } else {
-                        eprintln!("You need to enter a value of no more than 300 TERAGAS")
-                    }
-                }
-                Err(err) => return Err(color_eyre::Report::msg(err)),
-            }
-        };
-        Ok(Some(gas))
-    }
-
-    fn input_deposit(
-        _context: &super::TokensCommandsContext,
-    ) -> color_eyre::eyre::Result<Option<crate::types::near_token::NearToken>> {
-        eprintln!();
-        match crate::types::near_token::NearToken::from_str(
-            &Text::new(
-                "Enter deposit for a function call (example: 10NEAR or 0.5near or 10000yoctonear):",
-            )
-            .with_initial_value("1 yoctoNEAR")
-            .prompt()?,
-        ) {
-            Ok(deposit) => Ok(Some(deposit)),
-            Err(err) => Err(color_eyre::Report::msg(err)),
-        }
     }
 }
